// src/app/models/User_Setting_Model.ts
export class UserSettingModel {
    id:number;
    SportsName: string;
    SportsId: number;
    ComType: string;
    MatchCom: number;
    SessinCom: number;
    OddsMax: number;
    OddsMin: number;
    StakeMax: number;
    StakeMin: number;
    SessionMaxStake: number;
    SessionMinStake: number;
    MaxProfit: number;
    BetDelay: number;
    FancyDelay: number;
}